+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 Initial release
+
+## 0.0.2
+
+Add `Document` class to get info about a file (name, content uri, type, modification time, size,)
+Add `getDocumentTree()` method to list folder files by folder's content uri